Make TypedImplTrait its own thing.
