+++ sway-core/src/semantic_analysis/ast_node/mode.rs
@@ -1,0 +1,11 @@
+#[derive(Clone, Copy, PartialEq, Eq)]
+pub enum Mode {
+    ImplAbiFn,
+    NonAbi,
+}
+
+impl Default for Mode {
+    fn default() -> Self {
+        Mode::NonAbi
+    }
+}
